fix(pci-object-storage): add ods modal overflow hack
ref: DTCORE-3001
