+++ packages/manager/apps/pci-object-storage/src/pages/objects/container/add-user/AddUser.page.tsx
@@ -6,7 +6,7 @@ import {
 } from '@ovhcloud/ods-components/react';
 import { useNotifications } from '@ovh-ux/manager-react-components';
 import { ODS_BUTTON_VARIANT, ODS_SPINNER_SIZE } from '@ovhcloud/ods-components';
-import { useEffect, useRef, useState } from 'react';
+import { useRef, useState } from 'react';
 import { Translation, useTranslation } from 'react-i18next';
 import { useNavigate, useParams, useSearchParams } from 'react-router-dom';
 import { ApiError } from '@ovh-ux/manager-core-api';
@@ -15,6 +15,7 @@ import { TUser } from '@/api/data/user';
 import StepOneComponent from './StepOne.component';
 import StepTwoComponent from './StepTwo.component';
 import { useAddUser, useAllStorages } from '@/api/hooks/useStorages';
+import { useOdsModalOverflowHack } from '@/hooks/useOdsModalOverflowHack';
 
 export default function AddUserToContainerPage() {
   const { t } = useTranslation('containers/add-user');
@@ -30,7 +31,7 @@ export default function AddUserToContainerPage() {
   const [searchPrams] = useSearchParams();
   const containerId = searchPrams.get('containerId');
 
-  const modalRef = useRef(undefined);
+  const modalRef = useRef<HTMLOdsModalElement>(undefined);
 
   const { data: storages, isPending: isStoragesPending } = useAllStorages(
     projectId,
@@ -93,25 +94,10 @@ export default function AddUserToContainerPage() {
     },
   });
 
-  const isPending = isPendingListUsers || isPendingAddUser || isStoragesPending;
+  // @TODO refactor when ods modal overflow is fixed
+  useOdsModalOverflowHack(modalRef);
 
-  /* The following useEffect is a hack to allow the OdsSelect overflowing
-   * in the OdsModal thus allowing the user to select an item even for long lists.
-   * This hack can be removed once this issue is fixed on ods side, and this
-   * version of ods is available in pci-common & manager-react-components.
-   */
-  useEffect(() => {
-    if (!modalRef.current) return;
-    const { shadowRoot } = modalRef.current;
-    if (!shadowRoot.querySelector('style')) {
-      const style = document.createElement('style');
-      style.innerHTML = `
-        .ods-modal__dialog { overflow: visible !important; }
-        .ods-modal__dialog__content { overflow: visible !important; }
-      `;
-      shadowRoot.appendChild(style);
-    }
-  }, [modalRef.current]);
+  const isPending = isPendingListUsers || isPendingAddUser || isStoragesPending;
 
   return (
     <OdsModal onOdsClose={onClose} ref={modalRef} isOpen>
